♻️ Make termbox2 muliplatform with wasm support.
WASM and Windows do not have termios support and will require much
different ways of interacting with the terminal.

This extracts a "platform " API that termbox can use so that all
platform specific details can be hidden behind a swappable facade.

The tricky part is that we need to preserve these key properties.

1. `termbox2.h` is a single header file that can be included into any
project as a unit. (no platform specific includes)
2. the individual source `.h` files and `.c` files can be worked upon
using standard IDE tools and language servers to check syntax,
navigate, and resolev symbols.

Ultimately, the consumer experience must be unchanged:

`#define TB_IMPL` + `#include <termbox2.h>` in one translation unit.
The assembled header is functionally identical to the original — same
API, same behavior, same single-file usage.

To accomplish this we re-organize the way that the header file is
built. `termbox2.h` is now created entirely from a template called `termbox2.h.in`

New source file layout:

  termbox2.h.in   — assembly template with BEGIN/END markers
  tb_api.h        — public API + implementation internals (#ifdef TB_IMPL)
  tb_platform.h   — platform interface declarations. Each platform
                    must implement these functions
  termbox2.c      — core logic
  tb_posix.c      — POSIX backend (termios, pipes, signals, terminfo)
  tb_wasm.c       — WASM backend (delegates to __tb_host_* imports)
  assemble.awk    — assembles termbox2.h from the above

Every platform now as a chance at the very beginning to allocate a
data structure where all its state will be stored. This state is treated as fully opaque by
termbox and is stored at `global.platform`. It is passed as the first
argument to each platform call. For example, in `tb_posix.c` the
struct looks like:

```c
struct tb_posix {
    int ttyfd;
    int rfd;
    int wfd;
    int ttyfd_open;
    int resize_pipefd[2];
    struct termios orig_tios;
    int has_orig_tios;
    char *terminfo;
    size_t nterminfo;
};
```

The .c files include headers directly and compile standalone. This is
so that LSP/editor support (clangd, jump-to-definition, diagnostics)
works out of the box, but local include directives are stripped during assembly.

`make termbox2.h` reassembles the single-header deliverable. The awk script
inlines headers (stripping include guards), injects .c contents (stripping
preambles), and wraps everything in BEGIN/END markers for traceability.

`make terminfo` regenerates terminfo tables directly in tb_api.h, replacing
content between codegen markers.
